refactor: move code to activity
